Update poissondistribution.tex
I think from equation 1.2.2 we jumps to the assumption np = \lambda t is a bit out of a sudden. We can immediately see np is related to counting the people in n intervals (up to infinity) while p ->0 if the interval is small enough. However, I think how \lambda t is related to np needs further explanation. I think this assumption is related to Homogeneous Poisson point process  which can also be defined as a counting process (a type of stochastic process). Shall I add some explanations there?